When ellipsizing, check that maxWidth > 0
It is possible that this method can be called before the GUI is
completely constructed, in which case maxWidth will end up being 0 and
then the only text shown is an ellipsis (plus, Juce will throw an
assertion). So if there is no known width, just skip the loop and set
whatever string was passed in.